Add percent DID ownership to footer
