+++ src/components/Footer.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { Link } from 'react-router-dom'
 import EthereumAuthenticated from './common/EthereumAuthenticated'
 // import GasEstimate from './common/GasEstimate'
-import NumberDIDOwned from './common/NumberDIDOwned'
+import DIDOwnership from './common/DIDOwnership'
 import HasWeb3 from './common/HasWeb3'
 
 import { Container, Menu, Segment } from 'semantic-ui-react'
@@ -61,7 +61,7 @@ export default () => (
           Event Feed
         </Menu.Item>
         <Status />
-        <NumberDIDOwned />
+        <DIDOwnership />
         {/*<GasEstimate />*/}
         <HasWeb3 />
         <EthereumAuthenticated />
